pyproject.toml
[project]
name = "maestro"
version = "0.8.0"
description = "A multi-agent platform with the vision to facilitate deploy and run AI agents."
authors = [
    {name = "IBM"}
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "python-dotenv>=1.0.1",
    "openai>=1.76.2,<1.99",
    "pyyaml>=6.0.2",
    "jsonschema>=4.23.0",
    "docopt-ng>=0.9.0",
    "langchain-community>=0.3.16",
    "streamlit>=1.44.0",
    "streamlit-mermaid>=0.3.0",
    "psutil>=7.0.0",
    "openapi>=2.0.0",
    "openai-agents[litellm]>=0.0.14",
    "pycron>=3.1.2",
    "beeai-framework>=0.1.31",
    "tiktoken>=0.5.0",
    "slack_sdk>=3.35.0",
    "nest-asyncio>=1.6.0",
    "pydantic-ai[logfire]>=0.1.8,<1.0.0",
    "opik>=1.7.22",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "dspy>=2.6.27",
    "kubernetes>=33.1.0",
    "ddgs>=9.4.0",
    "fastmcp>=1.0",
    "jinja2>=3.1.6",
    "crewai>=0.134.0",
    "anyio>=4.10.0",
    "pytest-asyncio>=1.2.0",
    "pytest-tornasync>=0.6.0.post2",
    "pytest-trio>=0.8.0",
    "pytest-twisted>=1.14.3",
    "twisted>=25.5.0",
    "argparse>=1.4.0",
    "pandas>=2.0.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.0",
    "ruff>=0.12.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.14.0",
    "pre-commit>=4.2.0"
]

[project.scripts]
maestro = "maestro.cli.run_maestro:__run_cli"

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["maestro*"]

[tool.setuptools.package-data]
"maestro" = ["images/*.png", "schemas/*.json"]

[tool.pytest.ini_options]
addopts = "-v -s --ignore=framework"
